Fix ETSAuto input dtype: mean/std must be float32

+++ etsauto_adapter.py
@@ -145,8 +145,8 @@ class ETSAutoLaneDetector:
         img = cv2.resize(img, (self.input_shape[1], self.input_shape[0]))
         # Normalize (ImageNet)
         img = img.astype(np.float32) / 255.0
-        mean = np.array([0.485, 0.456, 0.406])
-        std = np.array([0.229, 0.224, 0.225])
+        mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+        std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
         img = (img - mean) / std
         return np.expand_dims(img.transpose(2, 0, 1), axis=0)
